refactor: move db/server inline test blocks to sibling tests.rs (#862)
server/src/rate_limit.rs still had its unit tests as an inline
`mod tests { ... }` block; move them to server/src/rate_limit/tests.rs
per the convention db/src/books/tests.rs establishes. Pure move, no
logic/assertion changes.

db/src/ebook/parse.rs and server/src/security_headers.rs — the other
two db/server-crate files #862 lists — were already split this way by
earlier commits, so no change was needed there.

Partial progress on #862 (db/server crate subset only); frontend
instances are out of scope per the issue's own notes.
